% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapLintR-package.R
\docType{package}
\name{REDCapLintR-package}
\alias{REDCapLintR}
\alias{REDCapLintR-package}
\title{REDCapLintR: Tool for REDCap Dictionary Good Practices}
\description{
It examines a REDCap project's dictionary to identifies and reports potential problems.
}
\author{
\strong{Maintainer}: Will Beasley \email{wibeasley@hotmail.com} (\href{https://orcid.org/0000-0002-5613-5006}{ORCID})

}
\keyword{internal}
